+++ test/test_app.rb
@@ -13,13 +13,14 @@ class AppTest < Test::Unit::TestCase
   self.test_order = :defined
 
   def self.startup
+    News.destroy_all
+    Reaction.destroy_all
+
     news = NewsFetcher.new
     news.fetch
   end
 
   def self.shutdown
-    News.destroy_all
-    Reaction.destroy_all
   end
 
   def app
@@ -160,5 +161,30 @@ class AppTest < Test::Unit::TestCase
 
     parsed_body = JSON.parse(last_response.body)
     assert parsed_body.count > 0
+  end
+
+  # happy path
+  def test_post_content_view
+    user = User.first
+    params = {
+      user_id: user.identifier,
+      user_source: user.source,
+      news_id: News.first.news_id,
+      context_from: 'test'
+    }
+
+    post '/content-views/', params
+    assert last_response.ok?
+  end
+
+  # no user_id should return 400
+  def test_post_content_view_400
+    user = User.first
+
+    post '/content-views/',
+         source: user.source,
+         news_id: News.first.news_id
+
+    assert last_response.status == 400
   end
 end
